fv_requirements.txt
git+https://github.com/davidbau/baukit@main#egg=baukit
accelerate==1.2.1
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
attrs==25.1.0
bitsandbytes==0.45.3
bottleneck==1.3.6
certifi==2025.1.31
charset-normalizer==3.4.1
datasets==3.2.0
dill==0.3.8
docker-pycreds==0.4.0
filelock==3.17.0
fireducks==1.2.1
firefw==1.2.1
frozenlist==1.5.0
fsspec==2024.9.0
gitdb==4.0.12
gitpython==3.1.44
huggingface-hub==0.29.3
hydra-core==1.3.2
hydra-submitit-launcher==1.2.0
idna==3.10
jupyter==1.1.1
lightning-utilities==0.11.9
loguru==0.7.2
multidict==6.1.0
multiprocess==0.70.16
mypy==1.6.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.2.140
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
packaging==24.2
palettable==3.3.3
pandas==2.2.3
propcache==0.2.1
protobuf==5.29.3
pyarrow==19.0.0
pydantic==2.10.6
pydantic-core==2.27.2
pytest==8.3.5
python-dateutil==2.9.0.post0
pytz==2025.1
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scienceplots==2.1.1
scikit-learn==1.6.1
sentencepiece==0.2.0
sentry-sdk==2.22.0
setproctitle==1.3.5
six==1.17.0
smmap==5.0.2
submitit==1.5.0
tokenizers==0.21.0
torch==2.1.0
torchaudio==2.1.0
torchmetrics==1.6.1
torchvision==0.16.0
tqdm==4.67.1
transformers==4.50.0
triton==2.1.0
typing-extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uv==0.5.10
wandb==0.19.7
xxhash==3.5.0
yarl==1.18.3
